#179 create "Delete TC Confirmation"

+++ TrustCenterSearch.Presentation/ViewModel.cs
@@ -200,13 +200,19 @@ namespace TrustCenterSearch.Presentation
 
         private void DeleteTrustCenterFromHistoryCommandExecute(TrustCenterMetaInfo trustCenterToDelete)
         {
+            var deleteConfirmation = MessageBox.Show("Are you sure you want to delete this Trust Center?", "Delete Confirmation", MessageBoxButton.OKCancel,MessageBoxImage.Question);
+
+            if (!deleteConfirmation.Equals(MessageBoxResult.OK)) return;
+            
             this.UserInputIsEnabled = false;
+
             this.Core.DeleteTrustCenter(trustCenterToDelete);
 
             this.TrustCenterHistoryActive.Remove(trustCenterToDelete);
             this.TrustCenterHistoryInactive.Remove(trustCenterToDelete);
 
             this.CertificatesCollectionView.Refresh();
+            
             this.UserInputIsEnabled = true;
         }
